+++ src/common/logger/index.ts
@@ -1,0 +1,2 @@
+export { LoggerModule } from './logger.module';
+export { WinstonLoggerService } from './winston.logger.service';
